Move Moments types and errors to ports
