+++ packages/providers/src/ports/MomentsApiProvider/MomentsApiProvider.ts
@@ -1,4 +1,4 @@
-import { CreateMomentInput, CreateMomentResponse } from './types';
+import { CreateMomentInput, CreateMomentResponse, MediaStatus } from './types';
 
 /**
  * Provides methods for interacting with a moments API.
@@ -46,4 +46,11 @@ export interface MomentsApiProvider {
    * @returns {Promise<void>} A Promise that resolves when the media processing is complete
    */
   waitForMediaProcessing(mediaKey: string, timeOut?: number): Promise<void>;
+
+  /**
+   * Fetches the media processing status.
+   * @param {string} mediaKey - The key for the media file
+   * @returns {Promise<MediaStatus>} - A Promise that resolves with the media processing status
+   */
+  fetchMediaStatus(mediaKey: string): Promise<MediaStatus>;
 }
