+++ packages/providers/src/ports/MomentsApiProvider/errors/index.ts
@@ -1,0 +1,1 @@
+export { InvalidMediaError } from './InvalidMediaError';
