+++ packages/providers/src/ports/MomentsApiProvider/index.ts
@@ -1,2 +1,3 @@
+export { InvalidMediaError } from './errors';
 export { CreateMomentInput, CreateMomentResponse } from './types';
 export { MomentsApiProvider } from './MomentsApiProvider';
