+++ packages/providers/src/ports/MomentsApiProvider/types/index.ts
@@ -1,2 +1,3 @@
 export { CreateMomentInput } from './CreateMomentInput';
 export { CreateMomentResponse } from './CreateMomentResponse';
+export { MediaStatus } from './MediaStatus';
